+++ apps/web/app/page.tsx
@@ -1,4 +1,3 @@
-import { Banner } from "~/components/banner";
 import {
   ComponentsSection,
   ContributorsSection,
@@ -7,16 +6,16 @@ import {
   FigmaSection,
   HeroSection,
   HomeNavbar,
-  MainFooter,
   ReactSection,
   SocialProofSection,
   TailwindSection,
 } from "~/components/homepage";
+import { MainFooter } from "~/components/main-footer";
 
 export default function HomePage() {
   return (
     <div className="relative">
-      <Banner />
+      {/* <Banner /> */}
       <HomeNavbar />
       <main className="min-w-0 flex-auto divide-y dark:divide-gray-700">
         <HeroSection />