+++ apps/web/components/homepage/home-navbar.tsx
@@ -11,21 +11,14 @@ export function HomeNavbar() {
         root: {
           base: "sticky top-0 z-40 mx-auto flex w-full items-center justify-between border-b border-gray-200 bg-white py-0 text-gray-500 dark:border-gray-700 dark:bg-gray-800 dark:text-gray-400",
           inner: {
-            base: "mx-auto flex w-full max-w-none flex-wrap items-center justify-between px-4 py-2.5 xl:max-w-8xl xl:px-20",
+            base: "mx-auto flex w-full max-w-none flex-wrap items-center justify-between px-4 py-2.5 xl:max-w-8xl",
           },
         },
       }}
     >
       <div className="flex items-center gap-3">
-        <Link href="/" className="sr-only">
-          Flowbite React
-        </Link>
-        <Link
-          aria-hidden
-          href="/"
-          className="flex items-center gap-3 text-2xl font-semibold text-gray-900 dark:text-white"
-        >
-          <Image alt="" aria-hidden height="32" src="/favicon.svg" width="32" />
+        <Link href="/" className="flex items-center gap-3 text-2xl font-semibold text-gray-900 dark:text-white">
+          <Image alt="Flowbite React logo" src="/favicon.svg" width="32" height="32" className="size-8" />
           <span>Flowbite React</span>
         </Link>
         <div className="ml-4 hidden lg:flex">