+++ packages/ui/src/components/Datepicker/Datepicker.tsx
@@ -197,9 +197,9 @@ const DatepickerRender: ForwardRefRenderFunction<DatepickerRef, DatepickerProps>
   const getViewTitle = (): string => {
     switch (view) {
       case Views.Decades:
-        return `${startOfYearPeriod(viewDate, 100)} - ${startOfYearPeriod(viewDate, 100) + 90}`;
+        return `${startOfYearPeriod(viewDate, 100) - 10} - ${startOfYearPeriod(viewDate, 100) + 100}`;
       case Views.Years:
-        return `${startOfYearPeriod(viewDate, 10)} - ${startOfYearPeriod(viewDate, 10) + 9}`;
+        return `${startOfYearPeriod(viewDate, 10)} - ${startOfYearPeriod(viewDate, 10) + 11}`;
       case Views.Months:
         return getFormattedDate(language, viewDate, { year: "numeric" });
       case Views.Days: