+++ packages/ui/src/components/Datepicker/Views/Months.tsx
@@ -2,7 +2,7 @@ import type { FC } from "react";
 import { twMerge } from "tailwind-merge";
 import { mergeDeep } from "../../../helpers/merge-deep";
 import { useDatePickerContext } from "../DatepickerContext";
-import { getFormattedDate, isDateEqual, isDateInRange, Views } from "../helpers";
+import { getFormattedDate, isDateInRange, isMonthEqual, Views } from "../helpers";
 
 export interface FlowbiteDatepickerViewsMonthsTheme {
   items: {
@@ -36,11 +36,13 @@ export const DatepickerViewsMonth: FC<DatepickerViewsMonthsProps> = ({ theme: cu
   return (
     <div className={theme.items.base}>
       {[...Array(12)].map((_month, index) => {
-        const newDate = new Date(viewDate.getTime());
-        newDate.setMonth(index);
+        const newDate = new Date();
+        // setting day to 1 to avoid overflow issues
+        newDate.setMonth(index, 1);
+        newDate.setFullYear(viewDate.getFullYear());
         const month = getFormattedDate(language, newDate, { month: "short" });
 
-        const isSelected = isDateEqual(selectedDate, newDate);
+        const isSelected = isMonthEqual(selectedDate, newDate);
         const isDisabled = !isDateInRange(newDate, minDate, maxDate);
 
         return (