+++ packages/ui/src/helpers/generic-as-prop.ts
@@ -6,14 +6,12 @@ export type AsProp<C extends React.ElementType> = {
 
 export type PropsToOmit<C extends React.ElementType, P> = keyof (AsProp<C> & P);
 
-// eslint-disable-next-line @typescript-eslint/ban-types
-export type PolymorphicComponentProp<C extends React.ElementType, Props = {}> = React.PropsWithChildren<
+export type PolymorphicComponentProp<C extends React.ElementType, Props = object> = React.PropsWithChildren<
   Props & AsProp<C>
 > &
   Omit<React.ComponentPropsWithoutRef<C>, PropsToOmit<C, Props>>;
 
-// eslint-disable-next-line @typescript-eslint/ban-types
-export type PolymorphicComponentPropWithRef<C extends React.ElementType, Props = {}> = PolymorphicComponentProp<
+export type PolymorphicComponentPropWithRef<C extends React.ElementType, Props = object> = PolymorphicComponentProp<
   C,
   Props
 > & {