+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # create-flowbite-react
 
+## 1.0.8
+
+### Patch Changes
+
+- [#1433](https://github.com/themesberg/flowbite-react/pull/1433) [`a5d008e`](https://github.com/themesberg/flowbite-react/commit/a5d008eb62e216e7139ff26fc67a323bb6e54b32) Thanks [@SutuSebastian](https://github.com/SutuSebastian)! - add `AdonisJS` integration guide
+
+## 1.0.7
+
+### Patch Changes
+
+- [#1423](https://github.com/themesberg/flowbite-react/pull/1423) [`4350ffb`](https://github.com/themesberg/flowbite-react/commit/4350ffb524cdecd23162c84b4222204bf8910b2c) Thanks [@SutuSebastian](https://github.com/SutuSebastian)! - update packages
+
 ## 1.0.6
 
 ### Patch Changes